major refactor on renderer: creating SpriteBatcher class
