+++ gueepo2D/engine/CMakeLists.txt
@@ -42,6 +42,7 @@ set (SOURCES
     "core/renderer/Renderer.cpp"
     "core/renderer/RendererAPI.cpp"
     "core/renderer/Shader.cpp"
+    "core/renderer/SpriteBatcher.cpp"
     "core/renderer/Texture.cpp"
     "core/renderer/VertexArray.cpp"
     "core/renderer/VertexBuffer.cpp"
@@ -129,6 +130,7 @@ set(HEADERS
     "core/renderer/Renderer.h"
     "core/renderer/RendererAPI.h"
     "core/renderer/Shader.h"
+    "core/renderer/SpriteBatcher.h"
     "core/renderer/Texture.h"
     "core/renderer/VertexArray.h"
     "core/renderer/VertexBuffer.h"
