+++ gueepo2D/engine/utils/PixelBuffer.cpp
@@ -3,6 +3,7 @@
 
 #include "core/renderer/Renderer.h"
 #include "core/renderer/Texture.h"
+#include "core/renderer/SpriteBatcher.h"
 
 gueepo::PixelBuffer::~PixelBuffer() {
 	if (m_data != nullptr) {
@@ -42,7 +43,7 @@ bool gueepo::PixelBuffer::SetColor(uint32_t x, uint32_t y, uint8_t r, uint8_t g,
 
 void gueepo::PixelBuffer::Render() {
 	m_textureReference->SetData(m_data, m_width * m_height * sizeof(uint32_t));
-	Renderer::Draw(m_textureReference);
+	Renderer::s_spriteBatcher->Draw(m_textureReference);
 }
